+++ lib/omise/config.rb
@@ -57,8 +57,9 @@ module Omise
     #
     attr_accessor :user_agent_suffix
 
-    # Getter and setter to set your secret API key. This key can be found in
-    # your dashboard.
+    attr_accessor :app_key
+
+    # Backward compatibility with old API Keys naming conventions
     #
     # This key is used for all call to the main API.
     #